Give every locale its own RSS feed
There was one feed. It carried the default locale's posts, linked to
/blog/<slug> without a locale prefix, and declared <language> to match — yet
every page in every locale linked to it. A reader on a Dutch page was offered
a feed of English posts pointing at English pages, and nothing in the markup
said so.

The XML moves to lib/rss so both routes share it, and takes a locale: it
decides which posts are in the feed, which URLs the items point at (through
getPostUrl, so translated feeds link to translated pages) and what <language>
reports. /<locale>/rss.xml is added for the secondary locales, guarded by
getSecondaryLocales() — the same rule the locale-prefixed blog routes use.

getRssUrl(locale) joins the other URL helpers in lib/blog, and the two places
that linked the feed now go through it: the pill in FollowAlong and the
<link rel="alternate"> in BaseLayout.

Verified by building with i18n temporarily switched on: /nl/rss.xml is
emitted with <language>nl</language> and a self-link to itself, /rss.xml
still reports en and still links English posts, and the head link follows the
page — /blog points at /rss.xml, /nl/blog at /nl/rss.xml. With i18n off, only
/rss.xml is emitted, so single-locale builds are unchanged.
