+++ src/pages/rss.xml.ts
@@ -1,76 +1,19 @@
-import { getCollection } from 'astro:content';
 import type { APIContext } from 'astro';
-import siteConfig from '@/config/site.config';
-import { getPostSlug } from '@/lib/blog';
+import { buildRssFeed } from '@/lib/rss';
+import { getRssUrl } from '@/lib/blog';
 import { defaultLocale } from '@/i18n';
 
 /**
- * Escapes XML special characters
+ * The default locale's feed, at the site root. Secondary locales get their own
+ * at `/<locale>/rss.xml` — see `src/pages/[locale]/rss.xml.ts`. The XML itself
+ * lives in `lib/rss` so the two routes cannot drift.
  */
-function escapeXml(text: string): string {
-  return text
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&apos;');
-}
-
-/**
- * Formats a date to RFC-822 format for RSS
- */
-function formatRfc822Date(date: Date): string {
-  return date.toUTCString();
-}
-
 export async function GET(context: APIContext) {
-  // Get only the default locale's non-draft posts for RSS
-  const posts = await getCollection('blog', ({ data }) =>
-    data.locale === defaultLocale && !data.draft
-  );
-
-  // Sort posts by date (newest first)
-  const sortedPosts = posts.sort(
-    (a, b) => new Date(b.data.publishedAt).getTime() - new Date(a.data.publishedAt).getTime()
-  );
-
-  // Generate slug from post id (strip the default-locale folder prefix)
-  const getSlug = (id: string) => getPostSlug(id, defaultLocale);
-
-  const site = context.site?.toString() ?? siteConfig.url;
-  const siteUrl = site.endsWith('/') ? site.slice(0, -1) : site;
-
-  const items = sortedPosts
-    .map((post) => {
-      const link = `${siteUrl}/blog/${getSlug(post.id)}/`;
-      const categories = post.data.tags
-        .map((tag) => `<category>${escapeXml(tag)}</category>`)
-        .join('\n        ');
-
-      return `    <item>
-      <title>${escapeXml(post.data.title)}</title>
-      <link>${link}</link>
-      <guid>${link}</guid>
-      <description>${escapeXml(post.data.description)}</description>
-      <pubDate>${formatRfc822Date(post.data.publishedAt)}</pubDate>
-      <author>${escapeXml(post.data.author)}</author>
-      ${categories}
-    </item>`;
-    })
-    .join('\n');
-
-  const rss = `<?xml version="1.0" encoding="UTF-8"?>
-<rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
-  <channel>
-    <title>${escapeXml(siteConfig.name)}</title>
-    <description>${escapeXml(siteConfig.description)}</description>
-    <link>${siteUrl}</link>
-    <atom:link href="${siteUrl}/rss.xml" rel="self" type="application/rss+xml"/>
-    <language>${defaultLocale}</language>
-    <lastBuildDate>${formatRfc822Date(new Date())}</lastBuildDate>
-${items}
-  </channel>
-</rss>`;
+  const rss = await buildRssFeed({
+    locale: defaultLocale,
+    site: context.site?.toString(),
+    feedPath: getRssUrl(defaultLocale),
+  });
 
   return new Response(rss, {
     headers: {
